+++ common/src/main/java/com/faboslav/friendsandfoes/common/init/FriendsAndFoesEntityModelLayers.java
@@ -3,6 +3,7 @@ package com.faboslav.friendsandfoes.common.init;
 import com.faboslav.friendsandfoes.common.FriendsAndFoes;
 import com.faboslav.friendsandfoes.common.client.render.entity.model.*;
 import com.faboslav.friendsandfoes.common.events.client.RegisterEntityLayersEvent;
+import net.minecraft.client.model.animal.cow.BabyCowModel;
 import net.minecraft.client.model.animal.cow.CowModel;
 import net.minecraft.client.model.monster.illager.IllagerModel;
 import net.minecraft.client.model.geom.ModelLayerLocation;
@@ -36,13 +37,13 @@ public final class FriendsAndFoesEntityModelLayers
 	public static void registerEntityLayers(RegisterEntityLayersEvent event) {
 		//? if <= 1.21.8 {
 		/*event.register(COPPER_GOLEM_LAYER, CopperGolemEntityModel::getTexturedModelData);
-		*///?}
+		 *///?}
 		event.register(CRAB_LAYER, CrabEntityModel::getTexturedModelData);
 		//? if >=1.21.3 {
 		event.register(GLARE_LAYER, () -> GlareEntityModel.getTexturedModelData().apply(GlareEntityModel.ADULT_TRANSFORMER));
 		//?} else {
 		/*event.register(GLARE_LAYER, GlareEntityModel::getTexturedModelData);
-		*///?}
+		 *///?}
 		event.register(ICEOLOGER_LAYER, IllagerModel::createBodyLayer);
 		event.register(ICEOLOGER_ICE_CHUNK_LAYER, IceologerIceChunkModel::getTexturedModelData);
 		event.register(ILLUSIONER_LAYER, IllagerModel::createBodyLayer);
@@ -55,8 +56,13 @@ public final class FriendsAndFoesEntityModelLayers
 		//? if >=1.21.3 {
 		event.register(CRAB_BABY_LAYER, () -> CrabEntityModel.getTexturedModelData().apply(CrabEntityModel.BABY_TRANSFORMER));
 		event.register(GLARE_BABY_LAYER, () -> GlareEntityModel.getTexturedModelData().apply(GlareEntityModel.BABY_TRANSFORMER));
-		event.register(MOOBLOOM_BABY_LAYER, () -> CowModel.createBodyLayer().apply(CowModel.BABY_TRANSFORMER));
 		//?}
+
+		//? if >= 26.1 {
+		event.register(MOOBLOOM_BABY_LAYER, BabyCowModel::createBodyLayer);
+		//?} else if >= 1.21.3 {
+		/*event.register(MOOBLOOM_BABY_LAYER, () -> CowModel.createBodyLayer().apply(CowModel.BABY_TRANSFORMER));
+		*///?}
 	}
 
 	private FriendsAndFoesEntityModelLayers() {